% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{and_then}
\alias{and_then}
\title{And-then combinator}
\usage{
and_then(p1, p2, join_fun = paste0)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}

\item{join_fun}{The function to join the results from the two parsers.}
}
\description{
And-then combinator
}
\examples{
p3 <- and_then(literal("A"), literal("B"))
p3("ABCDE")   # success:  list("AB", "CDE")
p3("ACCDE")   #   fail :  list()

p3 <- and_then(literal("A"), literal("B"), list)
p3("ABCDE")   # success:  list(list("A", "B"), "CDE"))
p3("ACCDE")   #   fail :  list()
}
