% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{any_of}
\alias{any_of}
\title{Any-of combinator}
\usage{
any_of(char_vec0)
}
\arguments{
\item{char_vec0}{A vector of characters.}
}
\description{
Any-of combinator
}
\examples{
p6 <- any_of(c("A", "B", "C"))
p6("ABCDE")  # success:  list("A", "BCDE"))
p6("BBCDE")  # success:  list("B", "BCDE"))
p6("CBCDE")  # success:  list("C", "BCDE"))
p6("DBCDE")  #   fail :  list()
}
