% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators_recursion.R
\name{between}
\alias{between}
\title{#' High-order combinator
#' Choice combinator
#' @param list_of_parsers A list of parsers.
#' @examples
#' p5 <- choice(list(literal("A"), literal("B"), literal("C")))
#' p5("ABCDE")  # success:  list("A", "BCDE")
#' p5("BBCDE")  # success:  list("B", "BCDE"))
#' p5("CBCDE")  # success:  list("C", "BCDE"))
#' p5("DBCDE")  #   fail :  list())
#' @export
choice <- function(list_of_parsers) {
  Reduce(or_else, list_of_parsers)
}
#' Sequence combinator
#' @param list_of_parsers A list of parser.
#' @examples
#' p7 <- sequence(list(literal("A"), literal("B"), literal("C")))
#' p7("ABCDE")  # success:  list("ABC", "DE"))
#' p7("BBCDE")  #   fail :  list()
#' @export
sequence <- function(...) {
  Reduce(`%then%`, list(...))
}
Between combinator}
\usage{
between(p1, p2, p3)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}

\item{p3}{A parser.}
}
\description{
#' High-order combinator
#' Choice combinator
#' @param list_of_parsers A list of parsers.
#' @examples
#' p5 <- choice(list(literal("A"), literal("B"), literal("C")))
#' p5("ABCDE")  # success:  list("A", "BCDE")
#' p5("BBCDE")  # success:  list("B", "BCDE"))
#' p5("CBCDE")  # success:  list("C", "BCDE"))
#' p5("DBCDE")  #   fail :  list())
#' @export
choice <- function(list_of_parsers) {
  Reduce(or_else, list_of_parsers)
}
#' Sequence combinator
#' @param list_of_parsers A list of parser.
#' @examples
#' p7 <- sequence(list(literal("A"), literal("B"), literal("C")))
#' p7("ABCDE")  # success:  list("ABC", "DE"))
#' p7("BBCDE")  #   fail :  list()
#' @export
sequence <- function(...) {
  Reduce(`%then%`, list(...))
}
Between combinator
}
\examples{
p1 <- between(literal("("), string("hi!"), literal(")"))
p1("(hi!)")   # success:  list("hi!", "")
p1("BBCDE")   #   fail :  list()
}
