% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{choice}
\alias{choice}
\title{High-order combinator
Choice combinator}
\usage{
choice(list_of_parsers)
}
\arguments{
\item{list_of_parsers}{A list of parsers.}
}
\description{
High-order combinator
Choice combinator
}
\examples{
p5 <- choice(list(literal("A"), literal("B"), literal("C")))
p5("ABCDE")  # success:  list("A", "BCDE")
p5("BBCDE")  # success:  list("B", "BCDE"))
p5("CBCDE")  # success:  list("C", "BCDE"))
p5("DBCDE")  #   fail :  list())
}
