% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{\%alt\%}
\alias{\%alt\%}
\title{Or combinator (with multiple parsing results)}
\usage{
p1 \%alt\% p2
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
Success if any of the parsers returns success.
}
