% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators_recursion.R
\name{\%chain_by\%}
\alias{\%chain_by\%}
\title{Chained-by combinator}
\usage{
p \%chain_by\% op
}
\arguments{
\item{p}{A parser.}

\item{op}{A parser.}
}
\description{
Chained-by combinator
}
\examples{
p15 <- one_of(as.character(0:9)) \%chain_by\% literal("+")
p15("1+")        # success:  list("1", "+"))
p15("1+2+")      # success:  list("1+2", "+"))
p15("1+2+;")     # success:  list("1+2", "+;"))
p15("1+2+3+;")   # success:  list("1+2+3", "+;"))
p15("1;2+;")     # success:  list("1", ";2+;"))
p15("+1+2+3+;")  #   fail :  list()
}
