% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{\%using\%}
\alias{\%using\%}
\title{Constructor combinator}
\usage{
p \%using\% f
}
\arguments{
\item{p}{A parser.}

\item{f}{A function; typically a constructor of a datatype.}
}
\description{
Allows chaining with a constructor function that parses the
matched results into an AST
}
