% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{literal}
\alias{literal}
\title{Literal parser}
\usage{
literal(char0)
}
\arguments{
\item{char0}{A character.}
}
\description{
Make an single-character-matching parser
}
\examples{
parser0 <- literal("A")
parser0("ABC")  # succuess: list("A", "BC")

parser0 <- literal("B")
parser0("ABC")  #   fail : list()
}
