% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{memoise}
\alias{memoise}
\title{Memoisation of a parser}
\usage{
memoise(p)
}
\arguments{
\item{p}{A parser.}
}
\description{
This is used to tackle (infinite) left recursion in parsing,
and it is done by memoising the input so that if the input is called twice
with the same parser without evaluation, then an infinite recursion is
registered.
}
\examples{
\dontrun{
term_s <- literal("s")
term_s("s")
sum_s <- (sum_s \%then\% sum_s) \%|\% term_s
sum_s("ss")   # error: infinite left recursion
msum_s <- memoise(msum_s \%then\% msum_s) \%|\% term_s
msum_s("ss")
}
}
