% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators_recursion.R
\name{one_of}
\alias{one_of}
\title{One-of / Any-of combinator}
\usage{
one_of(cs)
}
\arguments{
\item{cs}{A character vector.}
}
\description{
Match any element in a vector.
}
\examples{
p6 <- one_of(c("A", "B", "C"))
p6("ABCDE")  # success:  list("A", "BCDE"))
p6("BBCDE")  # success:  list("B", "BCDE"))
p6("CBCDE")  # success:  list("C", "BCDE"))
p6("DBCDE")  #   fail :  list()
}
