% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{one_or_more}
\alias{one_or_more}
\title{One-or-more combinator}
\usage{
one_or_more(p)
}
\arguments{
\item{p}{A parser.}
}
\description{
One-or-more combinator
}
\examples{
p10 <- one_or_more(literal("A"))
p10("AAABC")  # success:  list("AAA", "BC"))
p10("BBCDE")  #   fail :  list()
}
