% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{or-combinator}
\alias{or-combinator}
\alias{\%|\%}
\title{Or combinator (with single parsing result)}
\usage{
p1 \%|\% p2
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
Success if any of the parsers returns success.
}
