% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{or_else}
\alias{or_else}
\title{Or-else combinator}
\usage{
or_else(p1, p2)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
Or-else combinator
}
\examples{
p4 <- or_else(literal("A"), literal("B"))
p4("ABCDE")  # success:  list("A", "BCDE"))
p4("BCCDE")  # success:  list("B", "CCDE"))
p4("CCCDE")  #   fail :  list()
}
