% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{p_map}
\alias{p_map}
\title{Map combinator}
\usage{
p_map(f, p1)
}
\arguments{
\item{f}{A function to be applied to the parsed output.}

\item{p1}{A parser.}
}
\description{
Map combinator
}
\examples{
parser0 <- literal("A")
add_prefix <- function(x) { paste0("PREFIX-", x) }
parser1 <- p_map(add_prefix, parser0)

parser1("ABCDE")  # success:  list("PREFIX-A", "BCDE")
parser1("BBDED")  #  fail  :  list()
}
