% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{report}
\alias{report}
\title{Report the input and output of a function}
\usage{
report(f)
}
\arguments{
\item{f}{A function.}
}
\description{
Report the input and output of a function
}
\examples{
report(max)(3, 4, 5)
}
