% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators_recursion.R
\name{sep_by}
\alias{sep_by}
\title{Separated-by combinator}
\usage{
sep_by(p, sep)
}
\arguments{
\item{p}{A parser.}

\item{sep}{A parser.}
}
\description{
Separated-by combinator
}
\examples{
p14 <- sep_by(one_of(as.character(0:9)), literal(","))
p14("1,")        # success:  list("1", ","))
p14("1,2,")      # success:  list("12", ","))
p14("1,2,;")     # success:  list("12", ",;"))
p14("1,2,3,;")   # success:  list("123", ",;"))
p14("1;2,;")     # success:  list("1", ";2,;"))
p14(",1,2,3,;")  #   fail :  list()
}
