% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{sep_by_1}
\alias{sep_by_1}
\title{Separated-by-at-least-one combinator}
\usage{
sep_by_1(p, sep)
}
\arguments{
\item{p}{A parser.}

\item{sep}{A character.}
}
\description{
Separated-by-at-least-one combinator
}
\examples{
p14 <- sep_by_1(any_of(as.character(0:9)), literal(","))
p14("1,")        # success:  list("1", ","))
p14("1,2,")      # success:  list("12", ","))
p14("1,2,;")     # success:  list("12", ",;"))
p14("1,2,3,;")   # success:  list("123", ",;"))
p14("1;2,;")     # success:  list("1", ";2,;"))
p14(",1,2,3,;")  #   fail :  list()
}
