% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{sequence}
\alias{sequence}
\title{Sequence combinator}
\usage{
sequence(list_of_parsers)
}
\arguments{
\item{list_of_parsers}{A list of parser.}
}
\description{
Sequence combinator
}
\examples{
p7 <- sequence(list(literal("A"), literal("B"), literal("C")))
p7("ABCDE")  # success:  list("ABC", "DE"))
p7("BBCDE")  #   fail :  list()
}
