% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{string}
\alias{string}
\title{String combinator}
\usage{
string(str0)
}
\arguments{
\item{str0}{A character string.}
}
\description{
String combinator
}
\examples{
p8 <- string("ABC")
p8("ABCDE")  # success:  list("ABC", "DE"))
p8("BBCDE")  #   fail :  list()
}
