% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_classes.R
\docType{data}
\name{symbs}
\alias{symbs}
\title{Symbols (Alpha-numeric and Punctuation)}
\format{An object of class \code{character} of length 93.}
\usage{
symbs
}
\description{
Symbols (Alpha-numeric and Punctuation)
}
\keyword{datasets}
