% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators.R
\name{then-combinator}
\alias{then-combinator}
\alias{\%then\%}
\alias{\%xthen\%}
\alias{\%thenx\%}
\title{And(-then) combinator and variants}
\usage{
p1 \%then\% p2

p1 \%xthen\% p2

p1 \%thenx\% p2
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
Success if both parsers return success.
}
\examples{
p12 <- string("extract") \%xthen\% string("THIS")
p12("extractTHISpart")  # success:  list("THIS", "part"))
p12("BBCDE")  #   fail :  list()
p12 <- string("keepthis") \%thenx\% string("IGNORE")
p12("keepthisIGNORE")          # success:  list("keepthis", ""))
p12("keepthisIGNOREcontinue")  # success:  list("keepthis", "continue"))
p12("BBCDE")  # fail :  list()
}
