% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{thenx}
\alias{thenx}
\title{ThenX combinator}
\usage{
thenx(p1, p2)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
ThenX combinator
}
\examples{
p12 <- thenx(string("keepthis"), string("IGNORE"))
p12("keepthisIGNORE")          # success:  list("keepthis", ""))
p12("keepthisIGNOREcontinue")  # success:  list("keepthis", "continue"))
p12("BBCDE")  # fail :  list()
}
