% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{xthen}
\alias{xthen}
\title{Xthen combinator}
\usage{
xthen(p1, p2)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}
}
\description{
Xthen combinator
}
\examples{
p12 <- xthen(string("extract"), string("THIS"))
p12("extractTHISpart")  # success:  list("THIS", "part"))
p12("BBCDE")  #   fail :  list()
}
