% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{zero_or_more}
\alias{zero_or_more}
\title{Zero-or-more combinator}
\usage{
zero_or_more(p)
}
\arguments{
\item{p}{A parser.}
}
\description{
Zero-or-more combinator
}
\examples{
p9 <- zero_or_more(literal("A"))
p9("AAABC")  # success:  list("AAA", "BC"))
p9("BBCDE")  # success:  list("", "BBCDE"))
}
