% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinators_recursion.R
\name{zero_or_one}
\alias{zero_or_one}
\title{Zero-or-one combinator}
\usage{
zero_or_one(p)
}
\arguments{
\item{p}{A parser.}
}
\description{
Zero-or-one combinator
}
\examples{
p11 <- zero_or_one(literal("A"))
p11("AAABC")  # success:  list("A", "AABC"))
p11("BBCDE")  # success:  list("", "BBCDE"))
}
