% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{between}
\alias{between}
\title{Between combinator}
\usage{
between(p1, p2, p3)
}
\arguments{
\item{p1}{A parser.}

\item{p2}{A parser.}

\item{p3}{A parser.}
}
\description{
Between combinator
}
\examples{
p0 <- zero_or_more(any_of(c(letters, " ", ".")))
p1 <- between(string("KEY-1"), p0, string("KEY-2"))

input_str <- "KEY-1 this is content. KEY-2"
p1(input_str)   # success:  list(" this is content. ", "")
p1("BBCDE")     #   fail :  list()
}
