% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{sep_by}
\alias{sep_by}
\title{Separated-by combinator}
\usage{
sep_by(p, sep)
}
\arguments{
\item{p}{A parser.}

\item{sep}{A character.}
}
\description{
Separated-by combinator
}
\examples{
p15 <- sep_by(any_of(as.character(0:9)), literal(","))
p15("1,")        # success:  list("1", ","))
p15("1,2,")      # success:  list("12", ","))
p15("1,2,;")     # success:  list("12", ",;"))
p15("1,2,3,;")   # success:  list("123", ",;"))
p15(",1,2,3,;")  # success:  list("", ",1,2,3,;"))
}
